requirements.txt
# Receipt Processing System Dependencies

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
PyMuPDF>=1.23.0

# OCR Libraries
pytesseract>=0.3.10
easyocr>=1.7.0

# Google Cloud AI Services
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0

# Web framework for chatbot
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Additional dependencies
requests>=2.31.0
pathlib2>=2.3.7
argparse>=1.4.0
logging>=0.4.9.6
datetime>=5.3
json5>=0.9.14
os-sys>=2.1.4
sys-platform>=1.2.0
typing-extensions>=4.8.0
pathlib>=1.0.1
subprocess32>=3.5.4

# Additional utilities (these are already included with the above packages)
# pathlib - built-in Python library
# typing - built-in Python library
# json - built-in Python library
# logging - built-in Python library
# sys - built-in Python library
# os - built-in Python library
# argparse - built-in Python library
# base64 - built-in Python library 